fix(transposition) handle falsey children of child streams
The current implementation does not handle the situation where
the root of a vTree has a child that is an instance of most.Stream
and that most.Stream has a value which has `undefined` for children
i.e. is a text element like h1, h2, etc.

This commit closes issue #9