feat() wrap top level vTree
>Diff and patch root element is not anymore a child of the
given container (to the DOM Driver). It is now the given# Please enter the commit message for your changes. Lines starting container itself. Diff and patch will not override
and destroy# with '#' will be ignored, and an empty message aborts the commit. the identity of the given container. Instead, it will preserve it.# On branch fix/issue8 In
fact, the given container element is parsed to a vtree, and is# Changes to be committed: used as the first vtree in the diff & patch sequence. This enables# modified:
src/index.js avoiding to rerender the same tree client-side if it was already# rendered server-side, which saves processing power especially on mobile devices. The semantics
of the root element Observable, get(':root'), for the DOM Driver has also changed: the first emission is what was given as the container. This is important for
attaching event listeners.

Building on the experience with cycle-dom @staltz has made an undeniable claim to why we should wrap the top level vTree

BREAKING CHANGE: wrap the top-level vTree with given root element

  before:

  container - <div id='app'></div>
  application - makeDOMDriver('#app')
  vTree - h('div', {}, [])
  renders - <div></div>

  after:

  container - <div id='app'></div>
  application - makeDOMDriver('#app')
  vTree - h('div', {}, [])
  renders - <div id='app'> <div></div> </div>

  The id and className of the root element given to makeDOMDriver() will be preserved
  and it will be given children from the application.

This is a feature to close issue #8