docs(README): add useful information
Add Travis-CI badge
Show how it can be installed via npm
Information on libraries which are used and exported.
Highlight the key differences between Motorcycle-DOM and Cycle-DOM
Add a basic usage example to show how it's imported and expected
to be used in a Motorcycle application.
Add information on API with links to exported variables respective
documentation.

Closes #15