feat(dom-driver): reuse event listeners
Split the DOM driver into it's own file, and separates the
exports into it's own file for a cleaner separation of concerns.
index.js Now imports functions which are not used elsewhere and
exports them for developer convience.
Add skipRepeatsWith(sameElements) to the stream passed to makeEventsSelector
to avoid recreating event listeners for DOM Nodes which have not changed.

References #4