feat(makeDOMDriver): pass a stream of the rootElem to makeElementSelector
This changes rootVNode$ to rootElement$ because the base of all of the
changes require working with the rootElement. view$ is wrapped up in
most.creat() again as well, and the reason for that is, without it, by the time
the vNode.elm has reached other functions it is still undefined.