fix(virtual-dom): handle text nodes that are childNodes of rootElement
When changing the structure of VNodes from having `.sel` to `.tagName`
`.className` and `.id` elements are much more likely to be reused.
Under this assumption Text Nodes that live under the root element
initially passed to makeDomDriver were not being patched or removed
because the API HTMLElement.children does not return Text Nodes.
The function `emptyNodeAt` also previously did not handle textContent
of nodes because of the misuse of `.children`. Using `.childNodes` fixes
this problem and required `api.tagName` to be updated to handle Text Nodes.