Tylors/feat/has css selector (#112)
feat(hasCssSelector): implement hasCssSelector function

Given a CSS selector **without** spaces, this function does not search children, it 
will return `true` if the given CSS selector matches that of the VNode and `false` 
if it does not. If a CSS selector **with** spaces is given it will throw an error.

This also implements a robuster `parseSelector`.